backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pymongo==4.10.1
pydantic==2.9.0
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
